requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
python-decouple==3.8
django-cors-headers==4.2.0
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.5.0